requirements.txt
scrapy
pymongo
python-dotenv
pmatplotlib
seaborn
pypyodbc
seaborn
Django==4.1.2
wordcloud
matplotlib
scikit-learn
ydata_profiling
category_encoders
django-crispy-forms
django-widget-tweaks
fastapi
pandas
joblib
catboost
